+++ demo/custom-data.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from "@angular/common/http";
 import { Subject } from "rxjs/Subject";
 
-import { CompleterData, CompleterItem } from "../src";
+import { CompleterData, CompleterItem } from "../src/ng2-completer";
 
 export class CustomData extends Subject<CompleterItem[]> implements CompleterData {
     constructor(private http: HttpClient) {
@@ -9,8 +9,8 @@ export class CustomData extends Subject<CompleterItem[]> implements CompleterDat
     }
     public search(term: string): void {
         this.http.get("http://mysafeinfo.com/api/data?list=seinfeldepisodes&format=json&nm=" + term + ",contains")
-            .map((data: Array<any>) => {
-                let matches = data.map((episode: any) => this.convertToItem(episode)).filter(episode => !!episode) as CompleterItem[];
+            .map(data => {
+                let matches = (<Array<any>>data).map((episode: any) => this.convertToItem(episode)).filter(episode => !!episode) as CompleterItem[];
                 this.next(matches);
             })
             .subscribe();