+++ src/services/local-data-factory.ts
@@ -1,0 +1,10 @@
+import { Injectable } from "@angular/core";
+
+import { LocalData } from "./local-data";
+
+@Injectable()
+export class LocalDataFactory {
+    public create() {
+        return new LocalData();
+    }
+}