+++ src/services/local-data.ts
@@ -1,11 +1,10 @@
-import { Injectable, EventEmitter } from "@angular/core";
+import { EventEmitter } from "@angular/core";
 import { Observable } from "rxjs/Observable";
 import { catchError } from "rxjs/operators";
 
 import { CompleterBaseData } from "./completer-base-data";
 import { CompleterItem } from "../components/completer-item";
 
-@Injectable()
 export class LocalData extends CompleterBaseData {
 
     public dataSourceChange: EventEmitter<void> = new EventEmitter<void>();
@@ -19,7 +18,7 @@ export class LocalData extends CompleterBaseData {
 
     public data(data: any[] | Observable<any[]>) {
         if (data instanceof Observable) {
-            const data$ = <Observable<any[]>>data;
+            const data$ = data as Observable<any[]>;
             data$
                 .pipe(catchError(() => []))
                 .subscribe((res) => {
@@ -30,7 +29,7 @@ export class LocalData extends CompleterBaseData {
                     this.dataSourceChange.emit();
                 });
         } else {
-            this._data = <any[]>data;
+            this._data = data;
         }
 
         this.dataSourceChange.emit();
@@ -43,7 +42,7 @@ export class LocalData extends CompleterBaseData {
             this.savedTerm = term;
         } else {
             this.savedTerm = null;
-            let matches: any[] = this.extractMatches(this._data, term);
+            const matches: any[] = this.extractMatches(this._data, term);
             this.next(this.processResults(matches));
         }
     }