+++ tests/services/completer-service.spec.ts
@@ -1,0 +1,13 @@
+import { inject, TestBed } from '@angular/core/testing';
+
+describe('CompleterService', () => {
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+
+        });
+    });
+
+    // it('should create local data');
+
+});