+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## 2.0.0 (2017-11-24)
+## 2.0.5 (2017-12-01)
+
+Library structure and build now uses [angular-library-starter](https://github.com/robisim74/angular-library-starter)
 
 ### Breaking changes
  - Angular version support changed to 5.x.x
- - Use rxjs 5.5 lettable operators
+ - Use rxjs 5.5.x lettable operators
 
-## 2.0.0-beta.1 (2017-09-24)
+## 2.0.0 (2017-09-24)
 
 ### Breaking changes
 - Use Angular 4 `HttpClient`